requirements.txt
# Force clean install
--no-cache-dir

# Core dependencies
python-telegram-bot[callback-data]==20.8
openai==2.6.1
faiss-cpu
pydub
python-dotenv
moviepy
PyMuPDF
python-docx
beautifulsoup4
requests
tqdm
